one build_layer function
